Add li-8 gamma results function.
Now prints the results for the 980 keV line of Li-8 as a measurement
and limit.  Because this is a very weak result, I didn't bother to
use MINOS.  (Also, the limits are not calculated rigorously anyway.)
Added geterr function and added comments to the getpar function to the
effect that it seems to be one-indexed, not zero-indexed... sigh.