Change gaussian handling of li-8 gammas
Add a function that returns a string giving a gaussian in a
format so that one of the parameters is the integral.  Fit
the Li-8 gammas using this instead of gaus(0).  Previously,
I had taken the result of gaus(0) and fed it into the full
fit as though it were the integral, when in fact it was the
norm.  I think that given the sigma, this accidentally was
close to right.