B-12 handling of accidental neutrons.
Now automatically find the accidental neutron rate for specifically
the set of cuts being used.

Add a parameter to the fit for this rate, and let it float with a
pull term that is the statistical error on the above determination.

Drop support for Masaki's cut variations.

Remove some commented out code.