Apply new neutron efficiency correction when not using near neutrons.
The choice to fold this into the dr efficiency is only an acceptable
design if we always use the dr efficiency.  Well, we mostly only do,
except for the full B-12 fit, where currently the technote documents
that we use any neutron, but there's a switch in the code to use
near ones only.